Support VPATH mode, ala GNUMake's VPATH
We use JMake in an environment where we depend on VPATH to allow support
for "sparse work areas" and "baselines".

A "baseline" is the area that contains all the source and compiled code
that makes up our application.  This is built regularly by our
release management group.

A "sparse work area" is an area where the
developer has only their updated files; all other files are found in the
baseline.

Our tools and build mechanics rely on the concept of VPATH to work
correctly.  This update is a zero-sum change for traditional JMake
use, but also allows for VPATH to be used, as desired.