Ensure that all nested classes are registered in the PDB.
The old logic for selecting only anonymous nested classes of the form
OuterClass$[number] was incorrect, as it assumed that such
classes are in numerical order in the InnerClasses section of
the class file, but that is not the case. As a result it would
miss some of those classes.