Fix stupid use of string for variable
This wasn't caught because the tests weren't run because the test file
was misnamed; this RB also fixes the name of the test file.

The tests would sometimes fail because jmake would miss an update, and
a fix for that is included here.  There's also a fix for a (probably
harmless but annoying) warning message.

Testing Done:
ant test

Also, manual testing along with
https://rbcommons.com/s/twitter/r/1825/

Reviewed at https://rbcommons.com/s/twitter/r/1824/